SourceForge patch #1121972 "Improve input encoding in epslatex terminal"
by Harald Harders <h.harders@tu-bs.de>.

Added the koi8-u encoding to epslatex.

Added \inputencoding{...} to the output file. This enables the user to
use different input encodings in the surrounding LaTeX file and in the
plot.

Added a user-friendly error message if the package graphicx.sty or
graphics.sty is not loaded.