SF patch [1234632] gnuplot history file without GNU readline
by Bastian Maerkisch.

This patch enables GNUPLOT_HISTORY if gnuplot is not compiled with GNU
readline. History is automatically saved on program exit and reloaded on start.
"set/unset history" commands are supported.

Other makefiles should be adjusted to allow GNUPLOT_HISTORY to be defined when
HAVE_LIBREADLINE is not.