SF RFE [1063878] Support for pre-compiled bgd.dll
by Bastian Maerkisch.

If gd.trm made use of the gdFontGetTiny() and related methods to read the font
pointers instead of referencing them as extern, gnuplot could more easily be
compiled to work with the pre-compiled GD library bgd.dll instead of having to
download and compile zlib, libpng and libgd.

* makefile.nt: Support for binary version of gdlib (bgd.dll).

* gd.trm: New BGD_EXPORT_DATA_PROT and gp_gdXXX() for Win32.