scan_palette_from_buf():
The read_input() status was ignored and consequently when a partial
read of the pipe happened a full line was not removed from pipe.  For
long palettes garbage remained in pipe when done and next command was
bogus.

read_input_line(): A new routine that will check status of read_input and
continue reading if only a partial read. If there are too many reads, the
routine gives warning.

Patch by Daniel Sebald.