Preliminaries for: draw_clip_arrow() operates on floating point values
This patch should be a no-op. It paves the way for an upcoming patch to update
the draw_clip_arrow() logic. Our code wasn't careful about off-by-one pixel
errors: when converting from floating-point terminal coords to integer terminal
coords, we would sometimes round and sometimes truncate. This patch is careful
to not change the behavior, so it should produce EXACTLY the same results as
before. I tested it by doing this:

  gnuplot/demo$ (cd ..; make -j3 ) && ../src/gnuplot -e 'set term post color' all.dem < /usr/bin/yes > /tmp/all.ps

The contents of /tmp/all.ps were identical before and after this patch, modulo a
few timestamp differences.

draw_clip_arrow() is more or less a wrapper to the arrow() terminal function.
This terminal function takes integer arguments in terminal coords, which is a
problem when drawing vector fields: short vectors (< about 1 pixel wide) can
either disappear or the arrowheads can point in a very wrong direction. This
patch series resolves these issues by keeping the floating-point values around
as long as possible, and invoking the existing arrow() terminal function in a
more careful way to make sure the resolution is preserved, even when the vectors
are short.