Synthetic commit inserted to rectify historical 4-0-stable branch.
Due to an error in 2004 the 4-0-stable branch was rooted in the wrong
place.  To make the branch tip content right, one change (to demo/gnuplot.rot 
needs to be inserted between the branch root and the historical first commit. 

Some sort of CVS vendor branch lossage was probably implicated here.