Add indentation rules for editors with EditorConfig
gnuplot source files use a variety of different indentation styles.
This leads to inadvertent mixtures of styles if the editor is not auto-
configured.  Of course sources could be changed manually to the default
of indent=4 and tab-size=8.  Instead, we make use of the fairly
widespread EditorConfig http://editorconfig.org/ configuration files.
These are recognized by many popular editors either automatically or via
easily available plugins.