print_line_with_error() must not alter global gp_input_line or inline_num
If called from int_error() or os_error() this would not matter because the
correct values will be reset anyway, but if called from int_warn() this
can truncate the current input line and/or print the wrong line number in
the warning message.  In particular it can cause premature exit from a
bracketed clause.
Bug #2007