term->waitforinput() must invoke libreadline rl_signal_handler() on ^Z
Gnuplot does not suspend immediately by Control-z when it is configured
with --with-readline=gnu; suspension is triggered by a subsequent input
character.

When SIGTSTP is received, the signal handler of readline sets a flag
and just returns. Routines of readline check this flag to invoke the
real handler to suspend. Since we replace rl_getc_function by our own
routine to catch input from mouse, we should also check this flag to
invoke the real handler. Othewise, we cannot suspend until next time
routines of readline are used.

Unfortunately, there is no portable ways to invoke the real handler of
readline at the moment. We need to use its internal routine. libreadline
will implement a public function for this purpose in a future release.

EAM: Original patch included the wrapper code directly in x11.trm;
I moved it into a separate routine wrap_readline_signal_handler()
so that it can be shared by other gnuplot terminals.