remove conditional compilation symbol EAM_OBJECTS
We are long past the initial experimental phase of supporting "set object".
The code implementing circles, rectangles, polygons, etc is now used for
plot styles and page layout elements as well as for individual objects.
It no longer makes sense to compile the program without this support, so
we can simplify the code by removing the conditional compilation directives.