Use search paths relative to gnuplot executable on MSDOS
On Windows, file paths for external files (helpfile, postscript, lua, js,
gnuplotrc) are interpreted relative to the location of the gnuplot
executable.  This patch implements the same for MSDOS.

If gnuplot is located in a directory named "bin", treatment is somewhat
different:  then a Unix-like directory layout is assumed and search paths
are treated relative to the top directory.