offline mousing terminals should not try to back-convert nonlinear axis coordinates
For the terminals that support offline mousing (svg canvas), conversion
from terminal [x,y] to plot coordinates by the javascript support routines
assumes a linear axis.  The core driver code should at least warn when this is
not the case.
Introduce a new convention for the existing output variables plot_logaxis_[x|y]
    =  1 this axis is log scaled (full support)
    =  0 this is a normal (linear) axis (full support)
    = -1 this is either a nonlinear axis or requires a custom format routine
         "set mouse mouseformat function <f(x,y)>"
Currently the 0 flag is used for linked x/x2 or y/y2 axes, which is
incorrect unless the linking functions are linear.   This case code be signalled
by a separate flag value but is hard to determine automatically.