Any changes by extend_primary_ticrange() affect both primary and secondary axes.
The routine extend_primary_ticrange() exists only to emulate the old
behaviour (i.e. prior to nonlinear axis code) of autoscaled logscaled axes.
If it extends the range on the primary axis, this change must also be
propagated to the secondary (user-visible) axis.  Otherwise the reported
axis min/max are not correct.  Fixes log-scale mousing in canvas terminal.