'set boxdepth <value>' causes splot boxes to be 3D shapes rather than silhouettes
By default 'splot with boxes' draws only a single rectangle whose width
defines an extent along x and whose height is z.  The new 'boxdepth'
parameter defines a non-zero extent along y. For boxdepth > 0 each box is
drawn using 5 rectangles: front back left right top.

TODO:  As it stands depth-order sorting of pm3d quadrangles mangles the
order of drawing because it uses the mean depth of the whole box.
For common viewing angles this tends to incorrectly render tall boxes as
if they were in front of short boxes.  This can be fixed by sorting on the
z-value of the base instead of the mean, but it will need a new pm3d option.