another try at timestamp.h, this time using build date
The Makefile rule for timestamp.h was broken when we switched from
cvs to git, since it relied on ChangeLog always being current.
Proposed replacement rules have so far failed to work for out-of-tree
builds or builds from a snapshot rather than a full git clone.
This attempt simply uses the current build date, which admittedly may
not be newer than the last-modified date but at least has the virtue
of not depending on git.