'set view map' failed to allow for nonlinear y axis (bugfix)
'set view map' and 'set pm3d map' both flip the direction of the y
axis so that the layout matches a 2D plot.  In the case of a nonlinear
y axis, in particular 'set log y', this means we have to flip the
linked primary axis as well.
Bug #2029