emulate logscale axis range extension of pre-v5 splot
The routine extend_primary_ticrange() emulates the old behaviour
(i.e. prior to nonlinear axis code) of autoscaled logscaled axes.
It extends the range of the plot to include the next tic at either end.
We were already calling it for 2D plots; now we call it for 3D plots as well.
Bug #2029