Provide an inline unicode->utf8 conversion routine ucs4toutf8()
Recently added support for unicode escape sequences relied on a call
into libfontconfig to convert the codepoint to a UTF-8 representation.
Now we add an in-line conversion routine ucs4toutf8() based on a stackoverflow
contribution by "Nominal Animal" and adapted by Bastian Märkisch.

If an escape sequence of the form \U+hhhh is encountered but the current
encoding is not UTF-8, preserve the full sequence including the leading
backslash so that there is a possibility of handling it later in individual
terminal drivers.