use signed rather than unsigned coordinates in do_arc()
This allows clipping of circles or arcs whose center is off-screen on the
left or botton (negative x or y coordinate).  Previously the wrap-around
of unsigned [cx,cy] caused these to be treated as very far off-screen in
the positive direction and thus clipping failed to notice that portions of
the arc could be visible.
Bug #2035