use tgamma() if available rather than exp(lgamma())
Because of a historical screwup in naming, the C language "gamma" function
was really ln(gamma).  Linux chose to provide this as "lgamma", but kept
"gamma" as an alias for compatibility.  Then BSD did the sensible thing and
used "gamma" to mean, well, gamma.  So much for compatibility.
Anyhow, we don't need to perpetuate the confusion internally in specfun.c.
Macro LGAMMA is lgamma, fallback 1: old-libm "gamma", fallback 2: hard-coded lngamma.
Macro TGAMMA is tgamma, fallback is gp_exp(LGAMMA)