Trap attempts to sprintf 64-bit ints with %d or %x or %i or %u
Gnuplot's own code knows to use "% PRId64" as an output format for
64-bit integer variables, but this doesn't protect against user-supplied
formats given to sprintf().  On little-endian machines printing a 64-bit
int with "%d" works so long as the value fits in 32 bits.  But on big-endian
machines the wrong 32 bits are used, yielding either 0 or -1.  To avoid this
we now check whether a "l" or "L" modifier is present before accepting an
integer variable output format spec.  If not, truncate the variable to 32 bits.
If the 32 and 64 bit versions express the same quantity (no overflow) then
print the 32-bit version with the user-supplied format.
On overflow print "NaN".

FIXME: Might be better to write "Inf" or invoke int_error() or modify the
user-supplied format to include the modifier prefix __PRI64_PREFIX.

Bug #2037