Do not "snap" sampled point of function domain to 0 for non-linear x axis.
The motivation for snapping a sample to be exactly zero was that f(0) may
be of special interest but coarse sampling that spans x=0 might not happen
to place a sample there.  However this makes no sense for nonlinear axes,
expecially log axes, because the sampling is actually being done on the
hidden axis not the visible one where the value is zero.
Bug #2040