./configure --without-tektronix to disable support for 1970s/80s era terminal emulators
Terminal types included (or not):
  tek410x tek40 vttek xterm kc_tek40 km_tek40 selanar sixeltek bitgraph

Notes:
Some of these terminal types also require ./configure --with-bitmap-terminals.
Unless you really need to use the tektronix graphics protocol for some reason,
the prefered option is to use gnuplot's sixelgd terminal and a terminal emulator
that supports sixel graphics. You do not need to configure in the tektronix
support for this use.