skip evaluation of using spec expression that depends on a missing value
The presence of a "missing" flag in an input data field is easy to detect
if it is encountered in place of a bare number ('using N'), but if it is
encountered during evaluation of an expression ('using (func($N))') then the
function evaluation may exit via int_error() before the missing value can be
flagged for the calling code.  Now we pre-screen expressions in a using spec
for the presence of "$N" and note the dependence on column N in a new field
use_spec.depends_on_column.  During data input, if this field is non-zero
the content can be checked for a missing value flag before trying to
evaluate the expression is was found in.
Bug #2042