reorganize the main switch statement in get_data()
The old organization was switch (# of columns) with each
case value 1, 2, 3, ... 7 having to handle all possible plot styles.
The new organization is switch (plot style) with each case style
LINES, VECTORS, etc handling data values as appropriate.
This happens to reduce the code size but the main motivation is to
make it easier to read the code and easier to add new plot styles or
plot options requiring additional columns in the using specifier.