Fix column header/title assignment for plot with histograms style errorbars
The "autotitle columnhead" and "title columnhead" options normally assume
that the column in question is the one in the 2nd using specifier.
Histograms are an exception to this because the "y" value is in the
1st using specifier not the 2nd.  Plain histograms (no errorbars) worked
anyhow because the column head is always limited to the total number of
using specs, which is one in this case.  But histogram errorbars require
two using specs, so the assignment failed.