Compress qt terminal canvas size by 1 pixel on y
The floating point coordinate space used internally by the terminal
is [0:ysize*oversample] but in terms of screen pixels it is [0:ysize-1].
Simply casting to (int) produces a bottom-of-screen y coordinate that
is one too large and hence is not drawn.  Now we define a constant
QT_YBASE that is subtracted from the float->int conversion so that the
bottommost row of pixels is in-range.  If you don't want this, predefine
QT_YBASE=0 before compiling.