Reconcile cairo terminal pixel coordinates and floating point coordinates.
The floating point coordinate space used internally by the terminal
is [0:size*oversample] but in terms of screen pixels it is [0:size-1].
Simply casting to (int) produces a bottom-of-screen y coordinate that
is one too large and hence is not drawn.  Similarly the right-of-screen
maximum x coordinate. Now we report one pixel smaller dimensions back
to the core code via term->xmax and term->ymax while retaining the full
canvas size internally to retain the oversampling boost in precision.
This affects all the cairo terminal variants including wxt.