remove option EXTENDED_COLOR_SPECS
This was only used by the vgagl terminal.  It required supporting
two parallel color assignment paths for pm3d, one of which was only
needed for vgagl but was used by all terminals if vgagl was configured in.
Due to bitrot this second path no longer works (probably) or at least does
not know about newer color options.  Removing this configuration option
breaks vgagl.trm. That is fixable at the cost of losing support for
gradient-filled quadrangles, which no other terminal has anyhow.