Replace macro STORE_AND_UPDATE_RANGE by a subroutine
This complicated macro was called from 50+ sites in the code.
I believe that the original logic for making it a macro was that the inline
code was faster than a subroutine call. I confirmed this by benchmarking
in 2008, but that was 10 years ago.   Using the current source on newer
hardware there is no appreciable difference in speed.  If anything, it may be
better to have one cached copy of the code rather than 50 uncached copies.