using cbticlabels() now take coordinates from the same column as color
EXPERIMENTAL feature (not previously documented).
The code was previously hardwired to use coordinate values from column 4
but this is not correct in general.  Now for 3D plots we take it from column
3 (always z) and for 2D plots we take it from the column in the last using spec.
Variable color is documented as always being drawn from the last using spec.

FIXME: This leaves at least one ambiguous case:
    splot 'foo' using 1:2:3:4:cbtic(5) with pm3d
This produces a 3D heat map in which column 3 is the z value and column 4 is
the color.  Previously the cbtics value would come from column 4; now it
comes from column 3.