Remove linux and vgagl terminals
Both of these terminals rely on gnuplot being able to write directly to the
system graphics device (/dev/vga or /dev/svga or /dev/console).
This is accomplished by calling into libraries libvga or libsvga, both
provided by the svgalib source. The gnuplot executable must be installed suid
so that it has sufficient privilege for this to work.  A corresponding linux
kernel module svgalib_helper must be available and loaded into the kernel.

This was reasonable 20 years ago; now, not so much.

Nothing in a typical modern linux system requires or uses svgalib and
most distros do not include it by default.  It is not clear that it
is even possible to build it from the available source to work with current
linux kernels.

Furthermore the gnuplot core color and pm3d routines had an alternative code
path controlled by #ifdef EXTENDED_COLOR_SPECS that affected all terminals if
the linux vgagl terminal was configured.  This code path has not been
maintained and no longer works in 5.3 so I removed it in a previous commit.
That means currently the vgagl terminal would not compile even if a
working svgalib were available.

The recommended replacement for these terminals if you want to run gnuplot
from the linux console is to use a console terminal, e.g. yaft, that supports
sixel graphics. Gnuplot sixel output through yaft requires no special
configuration of either gnuplot or yaft.  This recommendation is documented
in a new dummy file term/linux-vgagl.trm that is automatically included in
the user manual.