new command "exit status <val>"
This forces an immediate exit with a user-specified exit status.
EXPERIMENTAL because IMO the usefulness of this is problematic.
Users would logically expect that if there is an actual error this
will be indicated in the exit status, but in general it isn't.
So status 0 tells you nothing; there may or may not have been an error.
The new command is useful only in the context that there are multiple
normal exit conditions (not errors in the normal sense) and the user
will signal which one pertains using this command.