protect sequence of command line execution from compiler optimization
The main loop over shell command line arguments in plot.c uses the iteration
convention {argv++; argc--;}.  However if execution of a command line entry
triggers LONGJMP via int_error() then code reordering due to compiler optimization
may fail to track changes in argc.  This can result in reexecution of a previous
command line entry.  In an attempt to prevent this we copy argc to a variable
marked "static volatile" so that the compiler will not reorder code that uses it.