revise handling of waitforinput in monothreaded wxt terminal
The monothreaded variant of wxt was not returning control after
"pause mouse key" because of incorrect logic in the select() loop
in wxt_waitforinput().  The problem was easily demonstrated by running
mouselabels.dem.  NB: The fix may alter the conditions that lead to
an input character being lost when gnuplot is reading from a pipe.
If such a case is found, try "set debug 44" to see if this repairs it.