revise handling of waitforinput in x11 terminal
The x11 terminal was responding to window close events
(e.g. bind "Close" "exit gnuplot") and clearing any associated flags
but was not actually returning from X11_waitforinput.
This caused gnuplot to wait for the next keystroke on the keyboard
before continuing.  Now we treat event GE_reset as if it had been
received as a keypress event containing GP_cancel, which is another
path the program uses.