Remove broken support for CIE/XYZ color space
As reported by Risto A. Paju, the XYZ->RGB transformation in getcolor.c
is not correct. Furthermore it and has never been correct).
As with the YIQ color space removed a while ago, the transformation matrix
was apparently never checked and any script that relied on output from the
broken matrix never was correct.
Replace black->gold palette example with an RGB space equivalent.
Bug #2078