Support multiple named voxel grids
Syntax changes:
	set vgrid $gridname {size N}
	vclear $gridname
	unset $gridname

VOXELGRID is now a user variable type (i.e. parallel to INTGR/CMPLX/...)
Voxel grid names must begin with '$'.  As with datablocks, they are
tracked in the master list of user-defined variables.
This means that fit and plot2d must now distinguish between a
datablock name and a vgrid name.

TODO:
	'reset' does not free voxel grid storage. Maybe it should?