experimental code to produce dual-color surface (inside/outside) for splot $vgrid with isosurface ... fc linestyle <n>
This depends on the ordering of vertices in the triangle/quad
tables, which seems to be correct for the Heller triangle list
but there are some glitches in the quad+triangle table.
There are classes that really want 4 triangular facets but the
table format only allows for 3 facets (quadrangle or triangle).

TODO:
       Triggering by "fc ls <n>" is a convenience for debugging.
       The plan is to add an option 'set isosurface {offset <n>}'
       analogous to the 'offset' option for hidden3d.