clean up inside/outside color assignment for isosurfaces
"set isosurface {{no}insidecolor <n>}" enables or disables use of a separate
color for the inside of an isosurface. The offset <n> works the same way as
the offset for front/back coloring of hidden3d surfaces.  "noinsidecolor"
(n=0) makes the inside and outside surfaces the same.  n>0 increments the
basic linetype used to color the outside by n to generate a linetype for
the inside surface.  The default is n=1 (inside linetype 1 greater than
outside linetype), as it is for hidden3d.