preserve LC_NUMERIC locale inside 'using' clause of plot commands
In order to read input data in the user's preferred locale, the plot, splot,
and stats commands set the numeric locale prior to data input and reset it
afterwards.  Similarly in order to produce output in the preferred locale,
the sprintf and gprintf functions set/reset the numeric locale on entry/exit.
Unfortunately these two policies collide head-on if sprintf or gprintf
appears in a 'using' spec.  The set done by sprintf on entry is redundant and
the subsequent reset on exit cancels the numeric locale currently in effect
for data input.  As a consequence the next data input operation uses the
wrong numeric locale.

To resolve this we now wrap the set/reset_numeric_locale in sprintf and
gprintf with " if (!evaluate_inside_using) ".

Bug #2125