prevent extra read past the end of a datablock holding matrix data
Under some circumstances the data input loop for matrix data could
issue an extra read at the end of the data input stream.  For input
from a file this was harmless, as once the end of file is hit EOF is
returned for all subseqent operations.  However for input from a
datablock this extra read caused an illegal access to memory
immediately following the last datablock line.  Add a check for
previously having hit end-of-file.
Bug #2129