Clear STATS_* variables before performing 'stats' analysis
If the variables are not cleared, then quantities from a previous
analysis may be incorrectly interpreted as belonging to the most
recent one.  E.g. variables specific to a 2-column analysis were still
present but out of date following a subsequent 1-column analysis,
the STATS_columns variable was still present but out of date if a
file-based analysis was followed by a datablock matrix analysis.
Bug #2136

Note: variables from 'stats FILE using n name=columnheader(n)' are not
cleared because the column header contents are not yet known at that point.