qt: correct decomposition of keyPress events into <key> + <modifiers>
1) The qt terminal was returning event->text() rather than event->key().
However Qt5 (not sure about Qt4) has already applied modifier keys
Shift and Control to generate event->text() which meant that ctrl-A
was returned as 001 rather than as "A" + modifier <ctrl>.
Now we return event->key() itself except that lower case letters are
explicitly returned as lower case.
2) commit f47b24d0 changed the handling of <tab> key events to trigger
on keyRelease rather than on keyPress.  However it overlooked the need
to update the status of modifier keys when sending the event.

Bugs #2131 #2143