Consistently use search paths relative to gnuplot executable on OS/2
On Windows and MSDOS, file paths for external files (helpfile, postscript,
lua, js, gnuplotrc) are interpreted relative to the location of the
gnuplot executable.  This patch makes implements the same for OS/2.
This eliminates local platform specific code for helpfile and postscript.

If gnuplot is located in a directory named "bin", treatment is somewhat
different:  then a local Unix-like directory layout is assumed and
search paths are treated relative to the top directory.

The outboard driver gnupmdrv.exe is now also searched for in the directory
pointed to by the GNUPLOT_DRIVER_DIR environment variable (like
gnuplot_x11 and gnuplot_qt), or in the directory of the gnuplot executable.