disambiguate interpretation of `splot with labels`
'splot with labels' can mean either "plot from a list of labels provided
in column 4", in which case we want to call place_labels3d(), or
"auto-generate contour labels", in which case we want to call cntr3d_labels().
We should never call both of these for the same set of labels.
With this patch the program will assume that after `set contour` that labels
are contour labels unless modified by `nocontours`.
Bug #2154