terminal level implementation of linetype nodraw
The keyword "nodraw" can be provided in place of a linetype.
For major plot elements the core code notes this and skips drawing
altogether, e.g. "plot sin(x) lt nodraw" produces a blank plot for
all terminals.  However for many smaller details where a linetype
is permitted the core code makes no special check and leaves it for
the terminal to omit the affected line segments individually.
Some terminals already do this (qt lua) but most do not.
Some treat "linetype nodraw" as "linecolor background", which is
not the same, others ignore it as a linetype and apply whatever
color would otherwise be active.

This patch implements more thorough treatment of "lt nodraw" for the
cairo terminals, libgd terminals, and svg.