terminal level implementation of linetype nodraw (generic dumb)
The generic terminal routines null_dashtype() and null_set_color()
clobber the current linetype, which means that setting linetype
to LT_NODRAW would not survive the normal sequence of setting
linewidth/linetype/dashtype/color for terminals that use them.
Modify term_apply_lp_properties() to exit early in this case.
This makes it possible to support LT_NODRAW on terminals such as dumb
that use the generic support routines.