Apply label font to hypertext (svg)
Since svg hypertext is rendered asynchronously, the notion of "current font"
doesn't really apply.  Instead we explicitly store a hypertext font if one
is specified in the "plot" command, and have the javascript support routines
refer to it when rendering hypertext. In the current implementation this means
the font applies to all hypertext entries in the plot.