new option for polygons - 'set obj N depthorder'
This is an alternative to setting the drawing order to behind/back/front.
It includes the object in the list of depth-sorted of pm3d quadrangles
built by 'set pm3d depthorder'. It is useful only for quadrangle objects,
since that is the only thing sorted by pm3d.
The utility may be extended later to handle triangle objects or triangular
tesselation of larger polygons.