New LaTeX2e terminal pict2e
LaTeX2e provides a (working) pict2e package since about a decade as a
successor to the limited picture environment.  This eliminates most of the
shortcomings of the original picture environment.  In addition it can
fill polygons and is compatible with most TeX variants, like latex/dvips,
pdflatex, and xelatex. Note that color filling of boxes and polygons
requires the xcolor package, not the color package. Transparent solid fill
is supported with pdflatex and the transparent package.

Patch #771