avoid walking full list of labels when adding a new one
The time for insertion of a new label in a plot "with labels" was order
(N^2) due to walking through the entire list each time to find the end.
We can cut this time to zero by keeping a pointer to the end of the list.
 Labels		old (sec)	new (sec)
-----------------------------------------
   10^4		1.5	 	1.0
   10^5		160		7.3
   10^6	   	hopeless	 66

(cherry picked from commit 02f74814332cf04d95b3cac8da73a914da33224b)