New filter option "smooth zsort" sorts 2D points on z before plotting
Syntax
    plot FOO using x:y:z:color smooth zsort with points lc palette

The intended use is to filter presentation of 2D scatter plots with a
huge number of points so that the distribution of high-scoring points
remains evident. Sorting the points on z guarantees that points with
a high z-value will not be obscured by points with lower z-values.
Limited to 2D "plot with points".