gd.trm: Add alpha channel when drawing thick lines
In order to get decent lines with lw > 1 using gdlib we draw with a brush.
Now we give the brush an alpha channel, but as the brush size gets larger
the repeated implicit compositing operation as the line is drawn causes
the resulting line tends to come out too opaque unless the alpha value is
quite near to 1.  Nevertheless better than it was before.
Bug #2177