add conditional code in the "reset" command to debug save/load operation
The *.dem files in .../demo act as unit tests for many gnuplot features.
Each demo file ends with a "reset" command.  The conditional code intercepts
this command and interpose a save/load operation before the reset.
This way save/load is tested from a large number of starting states,
potentially revealing syntax errors in the output from "save".

One such error this found was sensitivity to the ordering of
link/unlink/log/unlog commands.  Now we only execute "unset log" if the
axis is currently logscale.