Ensure space between plot border and y axis label
When ytic labels are present the program allocates an extra 2 character widths
of space between the label text and the y axis label.  However after "unset ytics"
or "set ytics axis", there are no ytic labels to the left of the plot and that
2 character width blank space disappears, leaving the y axis label too close to
the plot border.
This patch re-inserts the same blank space if the ylabel is present but the
ytic labels are not.
This patch also tweaks the ylabel position by 1/4 character width in all conditions.
Bug #2181