consolidate PS_set_font and ENHPS_set_font into a single routine
New code handles commands like 'set tics font "/:Bold"' (note superfluous /).
It is reasonable that users would try this since it matches the enhanced
text syntax '{/:Bold text}' and in fact the documentation suggests this
is a legitimate use even though not all terminals handle it nicely.
This fix also requires a tweak to the utility routine stylefont().
Previous code failed on the last plot of the jitter, parallel, and
projection demos.