block redefinition of built-in functions
Previously this would issue a warning:
"udf shadowed by built-in function of the same name"
Due to shadowing the new definition was accepted but never used.
However accepting the redefinition could cause failure to evaluate
a real function reference used for assignment, as in:
  do for [i=1:10] { voxel(x,y,z) = 1 }
Now the redefinition is silently ignored.