Guarantee consistent canvas size used by pslatex-based terminals
The current value of "set size" is used when the terminal is initialized,
and [was] used also when the final \includegraphics[] was written and the
output file closed.  A mismatch between the two values led to mangled
rendering.   Now we save the size used during initialization and reuse it
when the terminal is closed, regardless of any changes to "set size" in the
interim.  Affects pslatex, epslatex, cairolatex terminals.
Bug #2197