Add support for object-like pixmaps
Syntax:
	set pixmap <index> "imagefile.png" {center} {front|behind}
	    at <position> width <x-extent>

The position and width may be given in any coordinate system.
By default the y-extent is calculated to retain the aspect ratio of
the original image. As with objects, once the pixmap is defined it
appears in all subsequent plots.  Uses libgd (required) for reading
png, jpeg, gif image files. Compile with -DPIXMAPS

TODO:
   Allow explicit y-extent