Bugfix: functions and an unbounded iteration in the same plot command
"plot" and "splot" both process a plot command in two passes, one for
data files and one for function evaluation.  Both passes need to step
through the full plot command, however only the data pass was correctly
allowing for empirical termination of an unbounded iteration.
The presence of an unbounded iteration caused the second pass, function
evaluation, to fail.  E.g.
      plot for [i=0:*] A index i, f(x)
The fix for "plot" is straightforward.  The fix for "splot" is incomplete
because it requires remembering the end point of the iteration during the
first pass so it can be recreated in the second pass.  I allocated only
a single status variable for this, so it will fail if there are multiple
unbounded iterations in the same splot command.
Bug #2201