reset defaults at end of unbounded iteration
splot (but not plot) recycles the current plot structure for use by
the next plot if it turns out to be empty, e.g. run off the end of an
unbounded iteration.  Not all the fields are reset to default values.
Fix this for "notitle"; additional resetting may be needed.
Bug #2201