2-sided pm3d coloring for 3D polygons
pm3d surfaces can have separate colors for top/bottom surfaces if
you 'set pm3d depthorder' and the fillcolor is given via a linestyle.
This is similar to the treatment of hidden3d surfaces.
Now we apply the same rule for polygons with property "depthorder"
(currently limited to triangles and quadrangles).