Extend pm3d polygon support to pentagons and larger
Introduce a separate list of polygon vertices for pm3d polygons.
struct quadrangle gets a "type" field and a union "vertex".
Vertices for triangles and quadrangles are stored as before directly in
struct quadrangle, in q->vertex.corners[4].
For pentagons and larger polygons the vertices are stored in a separate
array indexed by q->vertex.array_index.