Add complex function LambertW( z, k )
z is complex, integer k is the branch of the multi-valued function W_k(z)

Note that existing function lambertw(x) = real(LambertW( real(z), 0 )).
This implementation is based on example code by István Mező, further guided by
   R. M. Corless, G. H. Gonnet, D. E. G. Hare, D. J. Jeffrey, and D. E. Knuth,
   On the Lambert W function, Adv. Comput. Math. 5 (1996), no. 4, 329–359.
   DOI 10.1007/BF02124750.
The nominal accuracy is 1.E-13 but the method is unstable in regions where
the function is discontinuous.  I have tried to tweak the starting conditions
to avoid triggering this.