Use font information when estimating text extent
Primarily this is for the key title, but the mechanism could be used
in other places.

estimate.trm:  returns text width in characters units (as before)
               returns text height in 10ths of current font size
estimate_strlen: passes these through to caller
2D and 3D key box: use font-adjusted estimates to reserved space for key title

Note: The fractional font size calculation in estimate.trm assumes
that the terminal default font is 12pt, which of course is not necessarily
true but we have no general mechanism yet for storing or retrieving the
actual current font size.

Bug #2409